requirements.txt
discord.py
requests
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
spotipy
